% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSSOAMObjects.r
\name{addStrZeros.1}
\alias{addStrZeros.1}
\title{Insert structural zeros to prevent ties between houses when two nodes
have never lived in the same house at the same time.}
\usage{
addStrZeros.1(pNetList, pSIDHIDList)
}
\arguments{
\item{pNetList}{A list of social network matrices (typically for a set of
waves).
  * Class may be
    +  'matrix' (base R 2-dim array)
    +  'Matrix' (any of the s4 classes supported by pkg 'Matrix')
  * Rows and columns must be named by node identifier (usually SID)}

\item{pSIDHIDList}{A data.frame with ncol() == the number of waves + 1.
The first col is the participant ID, and the remaining cols are house
identifiers representing where the participant lived at each observation.}
}
\value{
Returns the list of network matrices in the same form they were
  input, but with structural zeros (10's) encoded for linkages between
  individuals in different houses.
}
\description{
Insert structural zeros to prevent ties between houses when two nodes
have never lived in the same house at the same time.
}
