% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSSOAMObjects.r
\name{addStrZeros.2}
\alias{addStrZeros.2}
\title{Insert structural zeros to prevent ties between houses.}
\usage{
addStrZeros.2(pNetList, pSIDHIDList)
}
\arguments{
\item{pNetList}{A list of social network matrices (typically for a set of
waves).
  * Class may be
    +  'matrix' (base R 2-dim array)
    +  'Matrix' (any of the s4 classes supported by pkg 'Matrix')
  * Rows and columns must be named by node identifier (usually SID)}

\item{pSIDHIDList}{A list of dataframes (or objects that can act like them,
such as data.tables), corresponding to the networks in pNetList. Each is
a lookup table that codes which nodes (SIDs) belong in which houses
(HIDs) so each has two columns (SID and HID), which index this
correspondence.}
}
\value{
Returns the list of network matrices in the same form they were
  input, but with structural zeros (10's) encoded for linkages between
  individuals in different houses.
}
\description{
Insert structural zeros to prevent ties between houses.
}
