% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSNetCreation.r
\name{assignRlpName}
\alias{assignRlpName}
\title{Converts a relationship "short name" string (e.g. FR, LO, etc. ) to a
full (and more descriptive) string (e.g. friend, loan, etc. )}
\usage{
assignRlpName(pRlpShortName)
}
\arguments{
\item{pRlpShortName}{The two-character 'short name' code for each
relationship type}
}
\value{
The name  as a string
}
\description{
Converts a relationship "short name" string (e.g. FR, LO, etc. ) to a
full (and more descriptive) string (e.g. friend, loan, etc. )
}
