% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSDescriptives.r
\name{btwCent}
\alias{btwCent}
\title{Calculate avg betweenness centrality (Freeman) for a network}
\usage{
btwCent(pNet)
}
\arguments{
\item{pNet}{('network') A 'network' class object (from pkg 'network').
Could be binary or non-binary.}
}
\value{
(numeric) A numeric variable that averages over the vector,
excluding any NAs.
}
\description{
Calculate avg betweenness centrality (Freeman) for a network
}
\details{
Wraps the sna function 'betwenness', which gives a vector of
the betweenness scores by node.
}
