% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSConChecksFixes.r
\name{checkForHousesOf1}
\alias{checkForHousesOf1}
\title{Check for houses with just 1 network member}
\usage{
checkForHousesOf1(rlpNet)
}
\arguments{
\item{rlpNet}{a tbl_df in 'longNet' or 'rlpNet format.
This table can contain any number of waves, but if there is more than
one wave, the check will be done within wave, i.e. so the user
can be sure that no single wave has a house with just one member surveyed.}
}
\value{
A tbl_df containing WID, and HID for any individuals &
   waves where the individual was the only one in his/her house.
}
\description{
Check for houses with just 1 network member
}
\details{
In such cases there is no 'network', so they should be excluded
   from any network analysis. Usually this should not arise but it can
   if individuals were excluded from a supposed-final analysis set because,
   e.g., they did not have enough observations (say, 2 or more across all
   waves) to be included.
}
\note{
This check is important to run before taking any network set into
   production analyses, because other data corrections can create houses
   of size 1 as a 'side effect' that may not otherwise be obvious.
}
