% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSConChecksFixes.r
\name{checkForSelfSelection}
\alias{checkForSelfSelection}
\title{Checks long-format data (from 'makeLongNet') for self-choices (loops)}
\usage{
checkForSelfSelection(rlpNet, pWave = 0)
}
\arguments{
\item{rlpNet}{(tbl_df)Input long-format, i.e. 'longNet' or 'rlpNet'
format.}

\item{pWave}{(scaler, default=0) The wave to select. If default, function
assumes rlpNet is for 1 wave only, & that is the wave to use. Otherwise
the function will throw an error.}
}
\value{
Outputs a tbl showing self-choices. Normally these should
   be deleted.
}
\description{
Self-choices are not allowed. RSiena ignores them, but they make graphs
   weird, and make the addition of structural zeros more difficult.
}
