% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSConChecksFixes.r
\name{checkInMultipleHouses}
\alias{checkInMultipleHouses}
\title{Check possible multiple house membership within 1 wave.}
\usage{
checkInMultipleHouses(rlpNet, pWave = 0)
}
\arguments{
\item{rlpNet}{(tbl_df) A tbl in 'longNet' or 'rlpNet' format. Can include
multiple waves but only one will be processed (see 'pWave' doc).}

\item{pWave}{(scaler, default=0) The wave to select. If default, function
assumes rlpNet is for 1 wave only, & that is the wave to use. Otherwise
the function will throw an error.}
}
\value{
A tbl_df of all unique SID-HID combinations where the SID
    appeared in more than one house as chooser, chosen, or both. The
    columns 'ego' and 'alt' are 1 or 0, respectively, depending on
    whether SID was a chooser (ego) and/or chosen (alt) in that
    particular house.
}
\description{
Check possible multiple house membership within 1 wave.
}
\details{
SIDs can be assigned to multiple houses in a given wave either
    because of an incorrect manual input to the database, or (more
    typically) because the SID actually was a member of multiple
    houses during the wave survey period, and ended up being surveyed
    in several houses (theoretically this shoudld not happen--RAs should
    check first if they've been surveyed since the wave began), or
    because they were on two or more house rosters (this should not
    happen either, actually; RAs should first that the current
    roster is accurate before each survey).
}
