% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSSOAMObjects.r
\name{CompLast}
\alias{CompLast}
\title{Returns a list of {M,m}atrices with the same dimensions as the input 'n'.
    For all row/col indices that are identified in 'x' as missing for a given
    time point, the entire row/col matching the index is set to the last
    known value.}
\usage{
CompLast(x, n)
}
\arguments{
\item{x}{A data frame with k + 1 columns, where k == length(n), and the
first column is an identifier column. Columns 2:ncol(x) must
correspond with matrices 1:length(n).}

\item{n}{A list of matrices (representing networks across time). Matrices
should be identical in dimensionality and should include row and column
names that match with the identifiers in the input x.}
}
\value{
A list of n matrices with the exact structure of the input 'n',
    in which rows corresponding with missing participants, as captured
    in the input 'x', set to all NA values.
}
\description{
Returns a list of {M,m}atrices with the same dimensions as the input 'n'.
    For all row/col indices that are identified in 'x' as missing for a given
    time point, the entire row/col matching the index is set to the last
    known value.
}
\examples{

}
