% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSUtilities.r
\name{DeltaNet}
\alias{DeltaNet}
\title{Given a list of networks representing time ordered observations of a
network, calculate aggregate change statistics from one observation to
the next.}
\usage{
DeltaNet(n)
}
\arguments{
\item{n}{A list of networks as, e.g., output by makeNetworkSet()}
}
\value{
A table characterizing network changes.
}
\description{
Given a list of networks representing time ordered observations of a
network, calculate aggregate change statistics from one observation to
the next.
}
