% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSConChecksFixes.r
\name{errChecks}
\alias{errChecks}
\title{Internal function to check common errors for these functions}
\usage{
errChecks(rlpNet, pWave)
}
\arguments{
\item{rlpNet}{(tbl_df) A network edge list in the form of rlpNet,
or longNet. May be multiwave, but only 1 wave is ever used.}

\item{pWave}{(scaler, default=0) The wave to select. If default, function
assumes rlpNet is for 1 wave only, & that is the wave to use. Otherwise
the function will throw an error.}
}
\description{
Internal function to check common errors for these functions
}
