% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSConChecksFixes.r
\name{fixHouse}
\alias{fixHouse}
\title{Remove selected network connections in particular houses}
\usage{
fixHouse(rlpNet, pWave, pSID.HID)
}
\arguments{
\item{rlpNet}{(tbl_df) A tbl in 'longNet' or 'rlpNet' format. Can include
multiple waves but only one will be processed (see 'pWave' doc).}

\item{pWave}{(scaler, default=0) The wave to select. If default, function
assumes rlpNet is for 1 wave only, & that is the wave to use. Otherwise
the function will throw an error.}

\item{pSID.HID}{(tbl_df) Each row represents an individual
(SID) and house (HID) for which any relationships involving that
individual and house should be deleted.Each row consists of two
columns, 'SID' and 'HID'. These names are assigned, regardless of
what they're called in this data table/frame, so it makes sense to
just call them that from the get-go.}
}
\value{
A tbl_df in the same format as the input data table. This table\
  will have have been purged of any network data observations with
  SIDs and HIDs that match a row in 'pSID.HID'.
}
\description{
There are instances where a study participant may have made relationship
choices (or received them) in multiple houses during one wave period, or
may have completed a survey in a house after all or most others in the
house did so, when the index individual was not listed on the house roster
and could not thus be rated by others. In the case of multiple houses, we
must decide which house to keep them in (usually the first
one they were surveyed in that wave), and eliminate any in or out-choices
in any other houses. In the case of incomplete responses from other house
members (e.g. because ego was not yet in the house when they all surveyed),
normally we eliminate this individual's data for analysis purposes.
}
