% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSUtilities.r
\name{getNameAsString}
\alias{getNameAsString}
\title{Returns a character string of the input variable's name}
\usage{
getNameAsString(pVarName)
}
\arguments{
\item{pVarName}{A variable or some named object.}
}
\value{
A character string of the name.
}
\description{
Returns a character string of the input variable's name
}
