% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSNetCreation.r
\name{getNet4OneRlp}
\alias{getNet4OneRlp}
\title{Returns an edgelist data.table for just 1 type of relationship}
\usage{
getNet4OneRlp(tbLong, pWavVec = c(1), pRlpName)
}
\arguments{
\item{tbLong}{A long-format tibble, output by 'makeLongNet'}

\item{pWavVec}{A vector of waves; default is just the 1st wave}

\item{pRlpName}{A character string with the name of the relationship;
this must be a name in dtLong corresponing to a type of relationship.
These are: friend, loan, help, conv, advc, rate.}
}
\value{
Returns an edgelist-format tibble, with the columns SID, HID,
    WID, and AltID for the selected relationship type. Will have the same
    number of rows as tbLong, less any for which the relationship selected
    for has NAs.
}
\description{
Returns an edgelist data.table for just 1 type of relationship
}
