% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSDescriptives.r
\name{makeCountByHouseByWave}
\alias{makeCountByHouseByWave}
\title{Make table showing number of residents x house x wave}
\usage{
makeCountByHouseByWave(longTB, includeAltID = FALSE, print2Cons = FALSE)
}
\arguments{
\item{longTB}{(tbl_df) A multiwave long-format (from 'makeLongNet')
tibble.}

\item{includeAltID}{(logical) If TRUE, include individuals who appear as
only altIDs in the counts (i.e. individuals who were residents at the
time of survey, but who did not participate in the survey)}

\item{print2Cons}{(logical) If TRUE, write the output to the console.}
}
\value{
(tbl_df) A tibble with cols HID, WID, Count, in order by HID, WID.
}
\description{
Make table showing number of residents x house x wave
}
\details{
Should be a method for the longTB.raw class, when there is one.
}
