% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSNetCreation.r
\name{makeListOfNetsByHouse}
\alias{makeListOfNetsByHouse}
\title{Makes a list of house-specific networks, for a single wave only.}
\usage{
makeListOfNetsByHouse(
  pInTB,
  pWav,
  pTypNet = "FR",
  pTypOut = "MX",
  pTHold = 0,
  includeAltID = FALSE
)
}
\arguments{
\item{pInTB}{A tibble of the network data in 'long' format, i.e., output
from the function 'makeLongNet'. Could be multiwave, but only one
wave will be used, based on 'pWav'.}

\item{pWav}{A (single) wave ID to select for.}

\item{pTypNet}{(character; default = "FR") A code for the relationship type
the network should represent.
Possibilities are:
  FR: friend (1=close friend,2=friend, 3=acquaintence, 4=stranger,
    5=adversary)
  LO: loan $ (1=$0, 2=$10, 3=$50, 4=$100, 5=$500)
  HE: help   (1=very likely, 2=likely, 3=maybe, 4=prob not, 5=no)
  CO: conversation (1=daily, 2=almost daily, 3=every few days, 4=weekly,
              5=almost never)
  AD: advice from (1=very often, 2=quite often, 3=regularly, 4=rarely,
               5=never)
  ST: strength (1=very strong, 2=strong, 3=weak, 4=none, 5=negative)}

\item{pTypOut}{(character; Default "MX") A code indicating format for
networks in the output list.
  "NT" => class 'network' (pkg network)
  "MX" => class 'matrix' (pkg 'Matrix').}

\item{pTHold}{A number from 1 to 5, or -1 to -5.
If positive, the number means 'this value or higher = 1; else 0'.
If negative, the number means 'this value or lower = 0; else 0'.
Default is 0, which will not do any recoding.}

\item{includeAltID}{(logical; default FALSE). If only survey participants
(choosers as well as chosen) are to be included, this parameter should
be FALSE. If relationship ratings given to non participants (Alters
only) are also to be included, set to TRUE. Note that if any house
network includes non-participant alters, this parameter MUST be
TRUE, or an error is thrown.}
}
\value{
If pTypOut = A list of tibbles, one for each house in the same format as
    pInTB.
}
\description{
The main purpose of this arrangement of the network data is to facilitate
house by house descriptive statistics.
}
