% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSNetCreation.r
\name{makeRegionVar}
\alias{makeRegionVar}
\title{Recodes a vector of House IDs into region IDs.}
\usage{
makeRegionVar(pHIDs)
}
\arguments{
\item{pHIDs}{A vector of House ID #'s (integers) to be assigned to regions}
}
\value{
A vector of region numbers the same length as the input vector
}
\description{
Region 1: NC \cr
Region 2: TX \cr
Region 3: OR
}
