% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSNetCreation.r
\name{makeSIDHIDLookups}
\alias{makeSIDHIDLookups}
\title{Make a w-length list of SID-HID lookup tables from a w-length list of
   edge-format tibbles (each of which, e.g., created by 'makeLongNet')}
\usage{
makeSIDHIDLookups(tbList, includeAltID = FALSE)
}
\arguments{
\item{tbList}{A list of tibbles, one for each wave, in the format created
by 'makeLongNet'. The data must have columns named SID and HID,
representing ego and house, AltID for alter (if 'includeAltID' = TRUE)
and WID for wave.}

\item{includeAltID}{Logical, default=FALSE. If set to TRUE, the
output lookup tables will include individuals who were chosen
as alters (in 'AltID' col), but may not have been egos (in
'SID' col), implying they did not survey that wave. Normally
this is not desired.}
}
\value{
If w is the number of distinct waves (WID), i.e. the length of
  the input list, a w-length list of SID-HID lookup data.tables is returned,
  showing SID membership in each house for each wave. Within a wave, any SID
  belong to one and only one HID.
}
\description{
Make a w-length list of SID-HID lookup tables from a w-length list of
   edge-format tibbles (each of which, e.g., created by 'makeLongNet')
}
\details{
Length of the input list must be >= 1. If includeAltID=FALSE,
  the resulting output lists will only include individuals who actually
  surveyed on the wave in question. Otherwise it may include some
  house residents who were nonparticipants.
}
\note{
Because an analysis cannot meaningfully include an individual who is
  in more than one house, the function will not work if such a condition is
  found. Thus, the input 'tbList' elements must already have been filtered
  to remove cases where an individual appears in more than one house.
}
