% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSNetCreation.r
\name{makeSIDInAnyWave}
\alias{makeSIDInAnyWave}
\title{Creates a tibble of SIDs x houses in each wave}
\usage{
makeSIDInAnyWave(pInLst)
}
\arguments{
\item{pInLst}{(list) A list of tbls, one for each wave of interest,
with columns SID and HID, e.g. as created by 'makeSIDHIDLookups'.}
}
\value{
Returns a single tbl (tbl_df) with one row for any of the SIDs
  found in pInLst, and 1 column for each wave, containing the HID the
  the individual was considered to have been living in for that wave.
}
\description{
Creates a tibble of SIDs x houses in each wave
}
\details{
If the individual was not a resident anywhere (i.e. not in the
  study), the value for that SID and wave is NA. If the individual switched
  houses while surveys were in progress, it is possible he/she could have
  had a longNet row from two houses; however, he/she could only have
  surveyed in one, and this is the one that should be assigned to that SID.
  If such a switch occurred but the individual did not survey in either
  house (we've never seen this but it's possible), then if that SID is to
  be included for some reason (normally one would not),
  then an ad hoc decision must be made in the data prep script.
}
