% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSDescriptives.r
\name{makeSurveyedRateTblByHouseByWave}
\alias{makeSurveyedRateTblByHouseByWave}
\title{Make table showing rates of participation (participants / eligible) x house
& wave}
\usage{
makeSurveyedRateTblByHouseByWave(
  longTBElig,
  longTBSurveyed,
  print2Cons = FALSE
)
}
\arguments{
\item{longTBElig}{(tbl_df) A tibble in the form created by 'makeLongNet';
for correct results, this table should include all individuals considered
survey-eligible for each wave, whether they took the survey or not.
IMPORTANT: The function assumes that this includes individuals showing up
only as alters!}

\item{longTBSurveyed}{(tbl_df) a tibble in the form created by 'makeLongNet'.
This input should normally include only individuals who show up as egos
in the data. IMPORTANT: the function assumes that alters-only in this
table should NOT be counted as 'participants'.}

\item{print2Cons}{(logical) If TRUE, the output is printed to the console.
(Note that this parameter prints only the whole rate table, not the
individual elig and surveyed tables)}
}
\value{
(tbl_df) A tibble with 5 columns: HID, WID, Elig, Surveyed, and
  'Rate' (ordered by WID within HID). 'Elig' shows the number of individuals
  per house per wave as counted by the union of SID and AltID in
  'longTBElig', and normally interpreted to mean, all survey-eligible
  individuals per house per wave. 'Surveyed' shows the number of individuals
  per wave as counted by SIDs only, in 'longTBSurveyed'.
}
\description{
Make table showing rates of participation (participants / eligible) x house
& wave
}
\note{
It is possible to use this function to calculate rates of anything by
  house by wave, if you want to, simply by how you filter the two input
  tables, bearing in mind how this function treats SIDs and AltIDs for each
  input.
}
