% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSSOAMObjects.r
\name{multiGroup}
\alias{multiGroup}
\title{Given a list of ID vectors and other arguments accepted by datToSiena(),
this function returns a sienaGroup object with groups defined by the
list of ID vectors supplied as the first argument.}
\usage{
multiGroup(idlist = NA, ...)
}
\description{
Given a list of ID vectors and other arguments accepted by datToSiena(),
this function returns a sienaGroup object with groups defined by the
list of ID vectors supplied as the first argument.
}
\author{
Nathan J. Doogan \email{doogan.1@osu.edu}
}
