% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSDescriptives.r
\name{netAPL}
\alias{netAPL}
\title{Calculate average path (geodesic) length for a network}
\usage{
netAPL(pNet)
}
\arguments{
\item{pNet}{('network') A 'network' class object (from pkg 'network').
Could be binary or non-binary.}
}
\value{
(numeric) A numeric variable that averages.
}
\description{
Calculate average path (geodesic) length for a network
}
\details{
Wraps the sna function 'geodist', which gives a matrix of
the pairwise distances, not the average. This function only includes
connected pairs.
}
