% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSDescriptives.r
\name{netDiam}
\alias{netDiam}
\title{Find the longest shortest path between any 2 nodes in a network}
\usage{
netDiam(pNet)
}
\arguments{
\item{pNet}{('network') A 'network' class object (from pkg 'network').
Could be binary or non-binary.}
}
\value{
(numeric) A numeric variable of the max value of this network,
ignoring all nodes that are not connected at all.
}
\description{
Find the longest shortest path between any 2 nodes in a network
}
\details{
Wraps the sna function 'geodist', which gives a matrix of
the shortest (directed) path between each pair of nodes.
}
