% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSDescriptives.r
\name{nIsolates}
\alias{nIsolates}
\title{output count of # of isolates in a network}
\usage{
nIsolates(pNet)
}
\arguments{
\item{pNet}{('network') A 'network' class object (from pkg 'network').
Could be binary or non-binary.}
}
\value{
(numeric) An integer count of the number of isolates (individuals
with no indegrees or outdegrees).
}
\description{
output count of # of isolates in a network
}
\details{
Wraps the sna function 'isolates', which gives a list of them,
not a count.
}
