% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSSOAMObjects.r
\name{QOLMixing}
\alias{QOLMixing}
\title{RSiena goodness-of-fit statistic function for a 3x3 mixing matrix of QOL
  Few details are supplied because this function is not intended
  to be used directly. Moreover, this is a specialized function for a
  specific project and a specific covariate.}
\usage{
QOLMixing(i, data, sims, wave, groupName, varName, levls = 1:9)
}
\description{
RSiena goodness-of-fit statistic function for a 3x3 mixing matrix of QOL
  Few details are supplied because this function is not intended
  to be used directly. Moreover, this is a specialized function for a
  specific project and a specific covariate.
}
