% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSNetCreation.r
\name{recodeRlpVar}
\alias{recodeRlpVar}
\title{Recodes a vector representing a relationship variable to 0-1}
\usage{
recodeRlpVar(pInVec, pTHold)
}
\arguments{
\item{pInVec}{A numeric vector representing some relationship variable}

\item{pTHold}{A number from 1 to 5, or -1 to -5. \cr
If positive, the number means 'this value or higher = 1; else 0'.\cr
If negative, the number means 'this value or lower = 0; else 0'.\cr
Default is 0, which will not do any recoding.}
}
\value{
Returns a numeric vector of 0's and 1's representing the recoded
   input vector.
}
\description{
Recodes a vector representing a relationship variable to 0-1
}
