% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSUtilities.r
\name{replaceNaNWithNA}
\alias{replaceNaNWithNA}
\title{Substitutes NA for NaN in a data.table}
\usage{
replaceNaNWithNA(pInTB)
}
\arguments{
\item{pInTB}{Any tibble or data.frame}
}
\value{
Returns the input tibble or df with NaNs replaced by NAs
}
\description{
If input is a data.frame, it will be converted to a data.table (which is
also a data frame)
}
