% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSConChecksFixes.r
\name{runConChecks}
\alias{runConChecks}
\title{Runs consistency checks on network data}
\usage{
runConChecks(rlpNet, pWave = 0)
}
\arguments{
\item{rlpNet}{(tbl_df) A tbl in 'longNet' or 'rlpNet' format. Can include
multiple waves but only one will be processed (see 'pWave' doc).}

\item{pWave}{(scaler, default=0) The wave to select. If default, function
assumes rlpNet is for 1 wave only, & that is the wave to use. Otherwise
the function will throw an error.}
}
\value{
A list of tbl_df's containing specifics needed to find
   and fix any inconsistencies.\cr
   Element 1: The SID and house ID of individuals found to be in multiple
      houses, along with two 0-1 variables 'ego' and 'alt', which are 1
      if ego has chosen alters in the house, and/or received choices in
      the house, respectively.\cr
   Element 2: The House ID, SID, and AltID of any rows in tbl_df that
      are duplicated\cr
   Element 3: Rows of the input tbl_df that are self-selections
      (loops)\cr
   In addition, console output is generated, summarizing the results
}
\description{
Runs consistency checks on network data
}
