% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSNetCreation.r
\name{s0Fill}
\alias{s0Fill}
\title{Finds SIDs of anyone in 2 or more different houses}
\usage{
s0Fill(netSet)
}
\arguments{
\item{netSet}{A 'networkSet' object, equivalent to the output of
the 'makeNetworkSet' function, i.e. a list structured as:
1. w 'matrix' objects of equal dimension n, with SIDs as
   row and column names;
2. One n x w+1 'tbl', sorted on the first column (SID) which gives
   + The SID of each individual in the entire network set as col 1;
   + Additional columns HID1, HID2, ...HID<w> giving the house
     membership ID for each one.}

\item{pInTB}{A tibble representing an edge-list multiwave set of networks,
  i.e., with at least variables SID, AltID, HID, WID. Any other vars
  are ignored. The output of 'makeLongNet' works for this, as does the
  object generated by 'getNet4OneRlp'

Fills a 'networkSet' of networks with structural zeros between houses}
}
\value{
A 'networkSet' object identical to the input network Set, but with
  structural zeros added (as explained in 'details')
}
\description{
Finds SIDs of anyone in 2 or more different houses
}
\details{
At each wave, the wth adjacency matrix has structural zeros (10)
  added for all pairs of individuals who are not in the same house for that
  wave.
}
