% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSNetCreation.r
\name{selectHouses}
\alias{selectHouses}
\title{Select or remove a specified set of houses from a network}
\usage{
selectHouses(rlpNet, pHIDVec)
}
\arguments{
\item{rlpNet}{(tbl_df) in in the form created by 'makeLongNet', and
meant to represent a network edge list. Will actually accept any such
input containing at least the variables SID, HID, WID, and AltID.
May be single- or multi-wave}

\item{pHIDVec}{(numerical vecto) Vector of House ID numbers. They
must be either all positive numbers, or all negative, or an error is
thrown. If all positive, these house IDs are to be included. If all
negative, these house IDs are to be excluded}
}
\value{
A tbl_df in the same format as the input, including only the
   user-specified set of houses from pHIDVec
}
\description{
Select or remove a specified set of houses from a network
}
