% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RHSSOAMObjects.r
\name{TriadCensus}
\alias{TriadCensus}
\title{RSiena goodness-of-fit statistic function for the 16-dimensional triad
  census. Few details are supplied because this function is not intended
  to be used directly.}
\usage{
TriadCensus(i, data, sims, wave, groupName, varName, levls = 1:16)
}
\description{
RSiena goodness-of-fit statistic function for the 16-dimensional triad
  census. Few details are supplied because this function is not intended
  to be used directly.
}
